+++ engine/include/Material.h
@@ -2,6 +2,7 @@
 #include "Texture.h"
 #include <glm/glm.hpp>
 #include <string>
+#include "Shader.h"
 
 struct Material {
   Texture *texture = nullptr;
@@ -9,4 +10,8 @@ struct Material {
   glm::vec2 textureScale = glm::vec2(1.0f);
   std::string vertShaderPath = "basic.vert";
   std::string fragShaderPath = "basic.frag";
+  Shader *shader = nullptr;
+
+  // texture and shader are pointers so that we can have
+  // multiple Materials share same shader
 };