+++ game/src/shaderStore.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <string>
+#include "Shader.h"
+
+void initShaders();
+Shader *getShader(const std::string &shaderName);